feat: exporter-alignment refinements for allocator validation
Alignment fixes discovered by the 25-seed end-to-end validation of the
isaac allocator exporter (all defaults preserve legacy behavior; the
existing pytest suite passes unchanged):
- cost.accounting: per_traveller (legacy default) | per_request
  (allocator-style: passenger components once per request, delay
  measured from the pickup departure)
- new optional max_earliness request column: the earliness budget of
  required_arrival_time decouples from max_travel_delay (-1 = unbounded)
- new optional vehicles.json cost_return_to_depot: the depot-return leg
  enters travel/distance/duration cost without being constraint-checked
- distance matrices load as float64 (int32 rounding could exceed the
  plan-cost tolerance); min_pause: 0 no longer resets the continuous
  drive counter at every stop
- README extended-format spec updated accordingly

+++ python/darpinstances/travel_time_provider.py
@@ -47,8 +47,10 @@ class MatrixTravelTimeProvider(TravelTimeProvider):
     """Travel time from a precomputed distance/travel-time matrix."""
 
     @classmethod
-    def from_csv(cls, path_to_dm: Path):
-        dm = pd.read_csv(path_to_dm, header=None, dtype=np.int32)
+    def from_csv(cls, path_to_dm: Path, dtype=np.int32):
+        # travel time matrices are integer seconds; distance matrices may carry
+        # full-precision floats (metres), so the dtype is overridable
+        dm = pd.read_csv(path_to_dm, header=None, dtype=dtype)
         return cls(dm.values)
 
     @classmethod
@@ -69,9 +71,9 @@ class MatrixTravelTimeProvider(TravelTimeProvider):
         return self.dm[from_index][to_index]
 
     @classmethod
-    def read_from_file(cls, dm_filepath: Path):
+    def read_from_file(cls, dm_filepath: Path, dtype=np.int32):
         if dm_filepath.suffix == '.csv':
-            return cls.from_csv(dm_filepath)
+            return cls.from_csv(dm_filepath, dtype=dtype)
         else:
             return cls.from_hdf(dm_filepath)
